解决：【ID1025672】【资源管理-编辑表】更新数据源点击确认后报错的问题

+++ data-assert-backend/src/main/java/com/hashtech/service/impl/TableSettingServiceImpl.java
@@ -156,7 +156,7 @@ public class TableSettingServiceImpl extends ServiceImpl<TableSettingMapper, Tab
             });
             AuthListResult auth = auths.getData().get(auths.getData().size()-1);
             authResult.setAllowCall(auth.getAuthAllowCall());
-            authResult.setAuthEffectiveTime(null!=auth.getAuthPeriodBegin() && auth.getAuthPeriodBegin().intValue()>-1?0:1);
+            authResult.setAuthEffectiveTime(null!=auth.getAuthPeriodBegin() && auth.getAuthPeriodBegin().longValue()>-1?0:1);
             authResult.setCallCountType(null!=auth.getAuthAllowCall()&&auth.getAuthAllowCall().intValue()>-1?0:1);
             authResult.setPeriodEnd(auth.getAuthPeroidEnd());
             authResult.setPeriodBegin(auth.getAuthPeriodBegin());
